+++ R/module-auth.R
@@ -18,7 +18,7 @@
 #' @name module-authentication
 #'
 #' @importFrom htmltools tagList tags singleton
-#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton uiOutput
+#' @importFrom shiny NS fluidRow column textInput passwordInput actionButton uiOutput 
 #'
 #' @example examples/module-auth.R
 auth_ui <- function(id, status = "primary", tags_top = NULL, 
@@ -60,30 +60,48 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
             class = paste0("panel panel-", status),
             tags$div(
               class = "panel-body",
-              if (!is.null(choose_language)){
-                choices = NULL
+              {
+                
+                choices = lan$get_language()
+                lan_registered <- lan$get_language_registered()
                 if(is.logical(choose_language) && choose_language){
-                  choices = lan$get_language_registered()
+                  choices = unname(lan$get_language_registered())
                 } else if(is.character(choose_language)){
-                  choices = unique(c(intersect(choose_language, lan$get_language_registered()), lan$get_language()))
+                  choices = unique(c(intersect(choose_language, unname(lan$get_language_registered())), lan$get_language()))
                 }
                 
-                if(length(choices) > 1){
-                  selected = ifelse(lan$get_language() %in% choices, 
-                                    lan$get_language(),
-                                    choices[1])
-                  
-                  tags$div(
-                    style = "text-align: left; font-size: 12px;",
-                    selectInput(
-                      inputId = ns("language"),
-                      label = NULL,
-                      choices = choices,
-                      selected = selected,
-                      width = "20%"
-                    )
-                  )
+                names(choices) <- choices
+                for(i in 1:length(choices)){
+                  ind <- which(lan_registered %in% choices[i])
+                  if(length(ind) > 0){
+                    names(choices)[i] <- names(lan_registered)[ind]
+                  }
                 }
+                selected = ifelse(lan$get_language() %in% choices, 
+                                  lan$get_language(),
+                                  choices[1])
+                if(length(choices) == 1){
+                  style = "display:none"
+                } else {
+                  style = "margin-bottom:-50px;"
+                }
+                tags$div(style = style,
+                         fluidRow(
+                           column(width = 4, offset = 4, uiOutput(ns("label_language"))),
+                           column(4,
+                                  tags$div(
+                                    style = "text-align: left; font-size: 12px;",
+                                    selectInput(
+                                      inputId = ns("language"),
+                                      label = NULL,
+                                      choices = choices,
+                                      selected = selected,
+                                      width = "100%"
+                                    )
+                                  )
+                           )
+                         )
+                )
               },
               tags$div(
                 style = "text-align: center;",
@@ -113,7 +131,7 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
                 sprintf("bindEnter('%s');", ns(""))
               ),
               tags$div(id = ns("result_auth")),
-              if (!is.null(tags_bottom)) tags$hr(), tags_bottom,
+              if (!is.null(tags_bottom)) tags$div(style = "margin-top:-10px;", tags$hr()), tags_bottom,
               uiOutput(ns("update_shinymanager_language"))
             )
           )
@@ -128,8 +146,14 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #' @param input,output,session Standard Shiny server arguments.
 #' @param check_credentials Function with two arguments (\code{user},
 #'  the username provided by the user and \code{password}, his/her password).
-#'  Must return \code{TRUE} or \code{FALSE}.
-#'  To use additionnals arguments, set them with \code{purrr::partial} (see examples).
+#'  Must return a \code{list} with at least 4 slots :
+#'  \itemize{
+#'   \item \strong{result} : logical, result of authentication.
+#'   \item \strong{expired} : logical, is user has expired ? Always \code{FALSE} if \code{db} doesn't have a \code{expire} column.
+#'   \item \strong{authorized} : logical, is user can access to his app ? Always \code{TRUE} if \code{db} doesn't have a \code{applications} column.
+#'   \item \strong{user_info} : the line in \code{db} corresponding to the user.
+#'  }
+#'  
 #' @param use_token Add a token in the URL to check authentication. Should not be used directly.
 #' @param lan An langauge object. Should not be used directly.
 #' 
@@ -147,7 +171,8 @@ auth_ui <- function(id, status = "primary", tags_top = NULL,
 #' @importFrom htmltools tags
 #' @importFrom shiny reactiveValues observeEvent removeUI updateQueryString insertUI is.reactive icon updateActionButton updateTextInput renderUI
 #' @importFrom stats setNames
-auth_server <- function(input, output, session, check_credentials, 
+auth_server <- function(input, output, session, 
+                        check_credentials, 
                         use_token = FALSE, lan = NULL) {
   
   ns <- session$ns
@@ -162,6 +187,7 @@ auth_server <- function(input, output, session, check_credentials,
       lan <- reactive(lan)
     }
   }
+  
   
   observe({
     session$sendCustomMessage(
@@ -188,8 +214,15 @@ auth_server <- function(input, output, session, check_credentials,
       output$update_shinymanager_language <- renderUI({
         shinymanager_language(lan()$get_language())
       })
+      
+      output$label_language <- renderUI({
+        tags$p(paste0(lan()$get("Language"), " :"), 
+               style = "text-align: right; font-style: italic; margin-top:5px")
+      })
+      
     }
   })
+  
   
   authentication <- reactiveValues(result = FALSE, user = NULL, user_info = NULL)
   
@@ -207,7 +240,7 @@ auth_server <- function(input, output, session, check_credentials,
       if (isTRUE(use_token)) {
         # add_token(token, as.list(res_auth$user_info))
         .tok$add(token, as.list(res_auth$user_info))
-        updateQueryString(queryString = paste0("?token=", token, "&language=", lan()$get_language()), session = session)
+        addAuthToQuery(session, token, lan()$get_language())
         session$reload()
       }