+++ dev/shiny-classic-sqlite/global.R
@@ -18,3 +18,7 @@ credentials <- data.frame(
 # if (!file.exists("credentials.sqlite")) {
   create_db(credentials_data = credentials, sqlite_path = "credentials.sqlite", passphrase = "supersecret")
 # }
+
+#conn <- DBI::dbConnect(RSQLite::SQLite(), dbname = "dev/shiny-classic-sqlite/credentials.sqlite")
+#read_db_decrypt(conn, passphrase = "supersecret")
+#DBI::dbDisconnect(conn)